feat(all): separate front-end program merging and export
