+++ src/front/mergers.rs
@@ -1,4 +1,3 @@
-mod convert;
-mod definition_table;
+pub mod definition_table;
 pub mod package;
 pub mod program;
