+++ src/lib.rs
@@ -23,7 +23,8 @@ mod tests {
 
         program_merger.read_package("pkg", mock_file_system);
 
-        let mut program = program_merger.export_program();
+        let front_program = program_merger.return_merged();
+        let mut program = front_program.export_program();
 
         optimize(&mut program, &mut vec![Box::new(DeleteUnused)]);
     }
